删除没有使用的变量，降低了部分变量的大小，解决了ADDR_OVERFLOW问题

+++ global_vars.c
@@ -1,6 +1,6 @@
 #include "headfile.h"
 
 uchar mode = DEBUGGING;
-uint seconds = 0;      // 本周期经过的秒数
-uint milliseconds = 0; // 毫秒数,为1000时令seconds++
-uint cycle = 180;      // 红绿灯运行一轮的周期
+ushort seconds = 0;      // 本周期经过的秒数
+ushort milliseconds = 0; // 毫秒数,为1000时令seconds++
+ushort cycle = 180;      // 红绿灯运行一轮的周期
